Implement order processing orchestrator , stock reservation , rollback mechanism and placeholder for retry payment logic

+++ Rp-ecommerce/src/main/java/com/ecommerce/Rp_ecommerce/payload/OrderItemDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public class OrderItemDTO {
 
     private Long orderItemId;
@@ -18,13 +20,13 @@ public class OrderItemDTO {
     private Integer quantity;
 
     @NotNull
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
-    private Double discount;
-    private Double finalAmount;
-    private Double shippingCharge;
+    private BigDecimal discount;
+    private BigDecimal finalAmount;
+    private BigDecimal shippingCharge;
 
     private String trackingId;
     private String deliveryPartner;
@@ -35,8 +37,8 @@ public class OrderItemDTO {
     }
 
     public OrderItemDTO(Long orderItemId, Long productId, String productName,
-                        Integer quantity, Double priceAtPurchase, Double totalPrice,
-                        Double discount, Double finalAmount, Double shippingCharge,
+                        Integer quantity, BigDecimal priceAtPurchase, BigDecimal totalPrice,
+                        BigDecimal discount, BigDecimal finalAmount, BigDecimal shippingCharge,
                         String trackingId, String deliveryPartner, String notes) {
         this.orderItemId = orderItemId;
         this.productId = productId;
@@ -86,43 +88,43 @@ public class OrderItemDTO {
         this.quantity = quantity;
     }
 
-    public Double getPriceAtPurchase() {
+    public BigDecimal getPriceAtPurchase() {
         return priceAtPurchase;
     }
 
-    public void setPriceAtPurchase(Double priceAtPurchase) {
+    public void setPriceAtPurchase(BigDecimal priceAtPurchase) {
         this.priceAtPurchase = priceAtPurchase;
     }
 
-    public Double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
-    public Double getDiscount() {
+    public BigDecimal getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Double discount) {
+    public void setDiscount(BigDecimal discount) {
         this.discount = discount;
     }
 
-    public Double getFinalAmount() {
+    public BigDecimal getFinalAmount() {
         return finalAmount;
     }
 
-    public void setFinalAmount(Double finalAmount) {
+    public void setFinalAmount(BigDecimal finalAmount) {
         this.finalAmount = finalAmount;
     }
 
-    public Double getShippingCharge() {
+    public BigDecimal getShippingCharge() {
         return shippingCharge;
     }
 
-    public void setShippingCharge(Double shippingCharge) {
+    public void setShippingCharge(BigDecimal shippingCharge) {
         this.shippingCharge = shippingCharge;
     }
